+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+### Changed
+
+- [#482](https://github.com/XenitAB/terraform-modules/pull/482) add support for non VirtualAppliance routes
+
 ## 2021.12.2
 
 ### Fixed